Further simplify mixin declaration
Requiring a `namespace` string in mixin creation clutters the method
signature and limits the user's options on how to attach the mixin chart
(i.e. it can only be attached to the chart instance directly). This API
leaves the method of reference attaching completely up to the user.

Not sure if this is a good idea: the user may not keep references to the
chart, which will be problematic if we later decide to implement more
mixin management methods (like `removeMixin`, for example... ooo! We
could call that `mixout`!)