Support CommonJS and fix UMD
I'm using this library with Browserify and found that it was unable to
detect d3 correctly.  This adds in the necessary `require` call to find
the d3 dependency.

I also changed the UMD signature to not expect that `define` and
`require` are globals in the sense of the runtime, but instead are
global within the encapsulating function scope (which may be the same as
the runtime's).  This will ensure d3.chart works as expected when
bundled.